feat(3) - Added top five users by article

+++ src/main/java/com/example/app/repository/TopUsersRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TopUsersRepository extends JpaRepository<NumberOfArticlesByUser, Long> {
+public interface TopUsersRepository extends JpaRepository<NumberOfArticlesByUser, Integer> {
     @Query(value = "SELECT u.email, count(u.email) AS number_of_articles FROM users u, article a WHERE u.id=a.user_id GROUP BY u.email ORDER BY count(u.email) desc limit 5", nativeQuery = true)
     List<NumberOfArticlesByUser> getTopFiveUsersByArticles();
 }
